libc: fix const val non-constant initializers + type mismatch suffix corrections (5241 -> 2536 errors)
- Change 'public const val' to 'public val' for 320 non-constant initializers
- Fix expected Int/actual ULong: remove uL suffix or add .toInt() (459 errors)
- Fix expected Int/actual UInt: remove u suffix or add .toInt() (48 errors)
- Fix expected Int/actual Long: add .toInt() (145 errors)
- Fix expected ULong/actual Int: add uL suffix (255 errors)
- Fix expected UInt/actual Int: add u suffix (231 errors)

+++ src/commonMain/kotlin/io/github/kotlinmania/libc/unix/linuxlike/linux/musl/b32/riscv32/Mod.kt
@@ -72,11 +72,11 @@ public data class MaxAlignT(
 
 public const val RLIM_INFINITY: RlimT = 0xFFFFFFFFFFFFFFFFuL
 public const val VEOF: ULong = 4uL
-public const val RLIMIT_RSS: RlimitResourceT = 5uL
-public const val RLIMIT_AS: RlimitResourceT = 9uL
-public const val RLIMIT_MEMLOCK: RlimitResourceT = 8uL
-public const val RLIMIT_NOFILE: RlimitResourceT = 7uL
-public const val RLIMIT_NPROC: RlimitResourceT = 6uL
+public const val RLIMIT_RSS: RlimitResourceT = 5.toInt()
+public const val RLIMIT_AS: RlimitResourceT = 9.toInt()
+public const val RLIMIT_MEMLOCK: RlimitResourceT = 8.toInt()
+public const val RLIMIT_NOFILE: RlimitResourceT = 7.toInt()
+public const val RLIMIT_NPROC: RlimitResourceT = 6.toInt()
 public const val O_APPEND: CInt = 1024
 public const val O_CREAT: CInt = 64
 public const val O_EXCL: CInt = 128
